Split voteServer.js from vote.js, fix up imports
