+++ packages/lesswrong/lib/collections/comments/callbacks.js
@@ -3,7 +3,7 @@ import { Posts } from "../posts";
 import { Comments } from './collection'
 import { addCallback, runCallbacksAsync, newMutation, editMutation, removeMutation, registerSetting, getSetting, Utils } from 'meteor/vulcan:core';
 import Users from "meteor/vulcan:users";
-import { performVoteServer } from '../../modules/vote.js';
+import { performVoteServer } from '../../../server/voteServer.js';
 import { createError } from 'apollo-errors';
 import Messages from '../messages/collection.js';
 import Conversations from '../conversations/collection.js';
